+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Update `Confluent.Kafka` library to the most up to date version `1.0.0-RC6`
+- Add `Consumer` function `consumeLastMessage`, which reads last message of the given topic, by the highest offset
 
 ## 5.0.0 - 2019-04-23
 - [**BC**] Add more common types